Finalize fix
Requirements to get it to work on Android:
 - Every interaction with Valora requires closing and opening Valora app
 - Account key must be backed up within Valora
 - Valora account must have a phone number

This diff:
 - Only have `?` in the URL if there exists query params
 - Close window if query params exist and we are able to parse. Close
 window in a React component, not in global scope
